Remove unnecessary file, refactor
