--- function/body_info.html
@@ -1,10 +1,0 @@
-<div class="preamble">
-    <div class="preamble__guild-icon-container">
-            <img class="preamble__guild-icon" src="{}" alt="Guild icon">
-    </div>
-    <div class="preamble__entries-container">
-        <div class="preamble__entry">{}</div>
-        <div class="preamble__entry">{} / {}</div>
-        <div class="preamble__entry preamble__entry--small"> Between {} and {}</div>
-    </div>
-</div>
